requirements.txt
Flask==3.1.0
flask_cors==5.0.1
geopy==2.4.1
numpy==2.2.3
psycopg==3.2.5
python-dotenv==1.0.1
scikit_learn==1.6.1
